Adding Safari extension source code files
This commit includes the source code files of the first extension
bundle of this project. This extension bundle is targeting the OS X
browser, Safari. Files added and committed are the main structure of
this bundle which includes all the icon images, global  HTML, CSS, JS
files, popover HTML, CSS, JS files, a injected JS file to inject
JavaScript to the browser’s web content and the bundle’s preference
list XML file.