Update for README.md file
In this commit, I had updated the README file and included the Icon of the project (not final), download links table section, a summary section, a story section, a task list section, a contributions section and the copyright and licensing sections. All this sections provide main information about the project as a whole.